move files to more conventional structure
