+++ tests/src/validateFilename.test.js
@@ -1,7 +1,7 @@
 'use strict';
 
 const tape = require('tape');
-const validateFilename = require('../../modules/validateFilename.js');
+const validateFilename = require('../../src/validateFilename.js');
 
 tape('First test', (t) => {
   // ...
